requirements.txt
torch>=2.0.0
torchvision>=0.15.0
faiss-cpu>=1.7.4
fastapi>=0.110.0
uvicorn>=0.23.0
streamlit>=1.30.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.66.0
python-multipart>=0.0.6
requests>=2.31.0
scikit-learn>=1.3.0
